refactoring: moved all sample data generators from TestRuns to a separate file
