+++ src/app.ts
@@ -11,7 +11,8 @@ import {TestRuns} from "./test-runs";
 // TestRuns.testExtractAllEnglishPages();   // WET file must already be downloaded!
 // TestRuns.createFilteredSampleDataForGroups3_4();
 // TestRuns.getWebsiteByURL();
-TestRuns.extractPagesByURL();
+import {SampleDataGenerator} from "./sample-data/sample-data-generator";
+SampleDataGenerator.extractPagesByURL();
 
 
 // These runs require the WET file to be already downloaded and unpacked:
